xyz_utilities Refactor
