+++ public/js/layer_cluster.js
@@ -1,7 +1,7 @@
 module.exports = { getLayer };
 
-const utils = require('./utils');
-const svg_symbols = require('./svg_symbols');
+const utils = require('./xyz_utilities/utils');
+const svg_symbols = require('./xyz_utilities/svg_symbols');
 
 function getLayer() {
