+++ public/js/layer_mvt.js
@@ -1,5 +1,5 @@
 require('leaflet.vectorgrid');
-const utils = require('./utils');
+const utils = require('./xyz_utilities/utils');
 
 function getLayer(){
