+++ public/js/locate.js
@@ -1,6 +1,6 @@
 const L = require('leaflet');
-const utils = require('./utils');
-const svg_symbols = require('./svg_symbols');
+const utils = require('./xyz_utilities/utils');
+const svg_symbols = require('./xyz_utilities/svg_symbols');
 
 module.exports = () => {
