+++ public/js/xyz_utilities/svg_symbols.js
